Add skeleton for player and server
This app uses [/app] and [/lib]. [/app] contains files that handle
command line related interfacing. [/lib] contains everything else that
is required for the app to function.

Currently, it does not have the implementation of a game. It can only
send message to players. This is just a work-in-progress to show that
the Rpcs are working. This should be changed in the future to implement
the actual gameplay loop.