Implement minimal interaction for gameplay loop
This commit adds some cards as a starting point, including the
following:
  - 5 powerless card types
  - Exploding kitten

Players are then allowed to send an action given a [Card.t list] hand,
as described by the Rpc. It is currently minimally implemented such that
the only valid [Action] is to draw.

Currently, the deck composition and starting hand size is hardcoded to
follow the original game closely.