Add action selection
To add action selection, [Logic] is a new module created for [client/].
It uses [Async.Reader] and [Async.Writer] to interact with IO
asynchronously. [Logic.get_input] and [Logic.get_action] are implemented
so that we repeatedly ask the client until a valid response.

Clients also no longer receive sexps from the server. Instead, they
should receive nicely formatted and human readable strings. This
resolves the TODO for [send_message].

Great deal must be put in to ensure that string parsing and encoding of
variants makes sense. In this commit, [Action.of_string] is made to be
case-insensitive, while [Card.to_string] are capitalised with Title
Case. [Rainbow_ralphing_cat] is a special case - refer to [Card].
Expect tests are added for the above.