Add [Hand] module
Using a [Card.t list] as a player's hand is getting out of hand. We need
to have a proper string representation of a hand so the player is able
to play the game effectively. Furthermore, we expect that validation
must be performed when playing a card from their hand. This would most
likely necessitate linting of a new [Hand] module.

This commit implements a [Hand.t] as a [int Card.Map.t] under the hood,
which requires us to derive [compare] and [include
Comparable.Make_binable].

A new test file [test_hand.ml] has been created as it now makes sense
for [Hand] to be tested by itself. Tests have been shifted there and a
new test has been added to check for the string representation of a
hand.