Add skip card
This commit adds a skip card. This card is playable and thus we add the
[Power] module, mirroring the [Powerless] module. This also requires us
to update [Outcome] to encompass possible played power cards as well as
[Action].

Unit tests are added to ensure that playing a skip card results in sane
outcomes, as well as the parsing of actions. This should be case
insensitive.

We also ensure that the server receives an action that is performable.
For example, if the client plays a Skip card without owning one, the
server will ask again.