Shift [Player] into its own file and add [name] field
Adding the name field is important for the future when logging is done
to alert players what other players are doing. It will also be useful
when the steal card is implemented in the future as there needs to be a
way to reference someone else. A TODO is added to ensure that this field
is used in the future.

An RPC is therefore added so the client can tell the server what its
name is.