Lint a new [Game_state] module and rename [Logic] to [Interaction]
To prepare for extension, [Logic] needs to split up to the code that
handles the state of the game and the input prompting with the player.

It uses the [core_kernel.nonempty_list] library and uses [Nonempty_list]
extensively to ensure a compile time guarantee that there are at least 2
players playing the game. The crux of the logic lies in [Game_state.t].
It is a variant type that guarantees that there is either a [Winner] or
the game is [Ongoing].

[Interaction] is a module that is only in charge of asking for input
from the player, and repeatedly asks for input until it is valid.

Other players need to get alerted as well, and [to_self_alert] prepares
for that in a future commit.