Add [See_the_future] power card
This commit adds the power card from the original game "See the future".
It allows the player to peek at the top 3 cards without changing the
order of the deck. The turn does not end after playing this card.

Notably, [Outcome] does not use [Played of Card.Power.t] anymore.
Instead, it is broken down to [Skipped] and [Saw_the_future of Card.t
list]. This is because actions may have additional information encoded
into the outcome. In the future, when power cards like [Steal] are
implemented, this may be complicated even further, and [Played] is not
easily extensible in its form.

The alert code is long because of handling English plurals (see card vs
cards).

During testing, deriving [enumerate] on [Saw_the_future of Card.t list]
no longer makes sense, as [all] of [Card.t list] is unbounded. Instead,
[all] has been changed to [all_mocked] under [For_testing], prompting
the tester to provide sane mocked values for the variants which take
arguments.

All tests have sane results and has been tested on the command line as
directed by [CONTRIBUTING.md].

A TODO is also added about the concern of testability of [Game_state].