Make [Game_state] purer and [Interaction] all impure actions
This commit makes [Game_state] not reliant on interacting with input and
output. It takes in callback functions, which can be supplied and
extended with the functions in [Interaction].

It now makes use of [Deferred.repeat_until_finished] more extensively,
which helps with readability, instead of making our own recursive
functions.

This also helps prepare for future commits where steal or defuse cards
are added, making actions that a player can take possibly not just be
drawing or playing. This makes it more extensible and understandable.

This should also help [Game_state] be more testable as it no longer
interacts with standard input and output by default. Since all of it is
supplied through the callback functions, it is easily mockable. However,
mocking an [Rpc.Connection.t] in [Player.t] seems to be non-trivial. A
TODO has hence been added.