Add defuse card and restructure logic in [Action]
This commit adds the defuse card. This is a relatively large commit as
the minimum changes required to add this logic is quite a lot.

The defuse card allows a player to re-insert the drawn exploding kitten
into the deck at any position of their choice without looking. This
means that an additional RPC, [Get_draw_or_play], needs to be created.
This RPC is only sent when the player has defused an exploding kitten.
This also prompts us to create [Deck.insert], which helps us with this
task.

This also means that [Action] is no longer always a [Draw] or [Play], as
previously assumed. [Action] is now possibly a [Draw_or_play] or
[Insert_exploding_kitten]. For completeness, it could also be a
[Pass_turn] or [Eliminate_player]. This resembles a finite state
machine, and streamlines the process more. [Game_state] does not need to
know the exact workings of the game, since it can ask
[Action.Next_step.of_outcome] for the next transition. [Next_step] is a
new module linted to encompass all of the above.

Hence, [Instant] also tracks what the [Next_step] is at all times now,
so it knows how to move on to the next state of the game.

[test_next_step.ml] is a new file that helps us assert that the
[Next_step.t] from an [Outcome.t] looks correct. Furthermore, the
original [test_action.ml] file can be split out into more fine-grained
files in charge of testing their own part of the program.