Allow players to start with 1 defuse by default
In accordance with the rules of the original game, each player starts by
owning 1 defuse card each, and then drawing 7 cards from the deck each.
Note that the number of players for this deck is 2 to 5.

This commit changes [default_without_exploding_kittens] to take in
[player_cnt] to determine the number of [Defuse]s to add. Furthermore,
error checking is done so [player_cnt] is of a supported number, in this
case, 2 to 5.