Add input hints to the client
It is currently not obvious that the player can type "Draw", "Skip", ...
or input an integer (positive to negative) when re-inserting an
exploding kitten.

This commit adds the [~input_hint] parameter to [Logic.get], which
is a string that is tagged on to the back of the [prompt].

Note that this does not perform any validation on whether the move is
playable, such as attempting to play a [Skip] when it is not in the
hand. The hints are merely what can be accepted by the server.