Add the shuffle card
According to the original game, the shuffle card is a power card that
can be played like see the future, allowing you to shuffle the deck when
playing it. This does not end your turn.

[Deck.shuffle] now takes in a labelled [deterministically] to enable
testing of the [Power Shuffle] card. We avoid optional arguments as
codebases are safer when every call of the function is forced to think
about whether the new optional argument makes sense or not. This
labelled parameter is propagated into all functions that make use of the
shuffle function.