Make [Deck.add_exploding_kittens] take [~player_cnt] instead
We should delegate the idea of adding 1 less [Exploding_kitten] than the
number of players in [deck.ml] instead of [game_state.ml]. This commit
does exactly that. This resolves the TODO that pertains to this issue.

Furthermore, tests are added to ensure that 1 less [Exploding_kitten] is
added than the [player_cnt], followed by a shuffle.