Lint [Without_exploding_kittens] module for code safety
This commit adds safety to the logic by creating a
[Deck.Without_exploding_kittens] module. When calling [default], it
gives you a pre-built deck without exploding kittens. It is expected
that the caller uses [add_exploding_kittens] before playing the game.

However, there are no checks enforcing this, and a forgetful coder could
accidentally start the game with no exploding kittens. This commit fixes
that, by introducing an intermediate [Without_exploding_kittens.t] type,
which can only be converted to a [Deck.t] via [add_exploding_kittens].

Miscellaneously, [draw_hand] is renamed to [deal].

[default] now takes in a [shuffled] parameter. This is so [default] can
be tested easily. A shuffled deck output is not very easy to decipher,
and hence makes sense for a parameter like this to be exposed. It is
possible to provide the function with the extra parameter only in
[For_testing] and keeping the signature the same in the [.mli]. However,
this parameter makes intuitive sense and improves code quality, and
hence was exposed in the mli.

Tests are added for both [default] and [deal]. Mistakes were
discovered via these tests. [deal] notably checks for [n < 0] for
badness instead of [n <= 0] now. [default] also checks for [player_cnt >
5] for badness instead of [player_cnt >= 5].