Add [Double] action
In the base game, playing 2 of any card allows you to steal a random
card from a player of your choice. The text / ability of the card is not
taken into account. This is what gives [Powerless] cards a useful
ability. This commit adds that action.

When adding [Double] to [Draw_or_play], it must encompass both the
[Card.t] and the target [Player_name.t]. This also means that deriving
[enumerate] and [string] is now impossible as [Player_name.t] is not a
variant type. Hence, my own [to_string] and [of_string] functions must
be defined. As a hint to the player on how to format the action,
[format_doc] is created and passed to the player during prompting.
Furthermore, [all] doesn't exist, so we have to use a similar
[all_mocked] pattern that is used in other parts of the codebase.

In [handle], [Double] must also be implemented, which makes us create a
new helper function [Player_hands.transfer_random_card]. We must also
define a new [Outcome] called [Stole_randomly] that tells us that the
action is successfully completed.

This prompts us to reconsider how to handle [Outcome]s. Now, there are 3
relevant parties to alerting. Previously, a personalised message for the
player of the card and the generic message for the rest of the players
sufficed. Now, a 3rd kind of alert, called a [specialised_alert], must
be created. This is for special players. In the case of [Stole_randomly
(card, target)], the target is a special player who deserves a
customised message. Furthermore, [on_outcome] is adjusted to use
[Turn_order.t] instead, where we created useful helper functions to
determine who to send these messages to.

Furthermore, extra care must be taken when stealing from a possibly
eliminated player. When a player is eliminated, [Player_hands] needs to
track and remember if the players hand is no longer interactable. In the
main game, upon elimination, the player discards their entire hand, so
they are not stealable from. Hence, [Player_hands] is now a map to
[Hand_or_eliminated.t] where it is either [Eliminated] or [Playing].
Then, every function checks for possible elimination. This is reflected
in [test_player_hands.ml].

Care is taken to ensure that [Player_name.t] is case-insensitive during
checking of equality.