Move [Next_step] and [Outcome] out into their own modules
[Action] is becoming too big of a god module, with too many
responsibilities. [Next_step] nor [Outcome] really have to do with
[Action] other than being returned and somewhat coupled in logic.
They should be split out into their own files. This commit does that -
merely a copy and paste movement into their own modules.

Miscellaneously, a TODO-soon for making a module that encapsulates turn
orders and alive players is added. This resolves the concern of
testability of [Game_state]. By breaking it down to sane chunks from the
library, there is no need to test it, as long as we test the turn order
module.

Another TODO-soon is making [Player_hands] encapsulate the deck as well.
The state of the deck and the players makes sense to be grouped
together, as a lot of operations involve these 2 objects and updating
them as a group. This is in things such as drawing hands, or dealing
hands.