Add [Turn_order] module
The logic in [Game_state] handles too much about turn order. This commit
introduces a new module that encapsulates all the logic of handling turn
orders. This will be handy when [Attack] is implemented which will mess
with the turn order a lot and have players take multiple turns.

Furthermore, this will help simplify [Game_state] and justify not
needing to test [Game_state], as every component of [Game_state] should
be individually tested.

The concept of [current_player], [waiting_players] and [spectators] are
remembered in a [Turn_order.t]. When players are eliminated, they go
into [spectators]. Furthermore, [waiting_players] is similar to
[other_players] in the original [Game_state] code. It has been given a
more descriptive name, since [spectators have entered the picture.