Make [Action.handle] use [player_hands] and [player_name]
In preparation to add actions that enable stealing of cards, it is no
longer sufficient for [handle] to only consider the [hand] of the
[current_player]. Instead, the signature must be changed to use
[player_hands], so it has context of the full game. Then, when a steal
is performed, the target is known and can be processed by the function.

Then, [Game_state] can be simplified as operations with [handle] no
longer deal with a [hand] but the full [player_hands] argument. The
[update] function in particular can be vastly simplified.

This does bring a downside, which is [Player_hands] being paranoid and
constantly threatening to return an [Error] because it can never be
certain that the supplied [Player_name.t] exists. This does not seem
like an issue that can be easily solved at compile time.