Add alerting for spectators
This commit adds [Outcome.to_uncensored_alert], which keeps dead players
in the information loop. Since dead players cannot interact with the
game anymore, all information can be sent to them without censoring. The
only caveat is that the messages must slightly be altered to make sense
to the spectators, such as using the player name instead of "You".

Since a spectator disconnecting should not crash the game,
[send_message] is also modified to take in a new parameter
[ignore_errors]. This ensures that even if a message cannot be sent to a
spectator because they disconnected, the game can still carry on.