Set ocamlformat to break at 80 columns
This commit only introduces one new setting in [.ocamlformat], [margin =
80], followed by running [dune fmt]. All comments are also manually
formatted to wrap at 80 characters.

This is an opinionated override on the default of 90 characters
introduced by [profile = janestreet]. 80 characters is a more widely
accepted and agreed upon limit to line lengths. When screens are more
cramped horizontally, such as during diffing, this setting shines
greater.