Add README.md and modify CONTRIBUTING.md
This commit adds a README.md file which helps new people to the
repository understand what it is about at a glance.

Furthermore, the CONTRIBUTING.md file has been updated to explain some
non-obvious choices in the repository. For instance, inline comments of
TODO-soon and TODO-someday, which are currently unexplained, as well as
the commit style of this repository.