Implement [Triple]
This commit implements the triple action. In the original game, playing
3 copies of the same card allows you to pick a player and name a card
that you want. If the player has that card, they must give it to you. If
they do not have that card, the action whiffs - you still lose your 3
cards and gain nothing.

The new outcomes are [Failed_to_steal_via_triple] and
[Stole_via_triple]. Notably, [Stole_randomly] has been renamed to
[Stole_randomly_via_double]. This is because between [Double], [Triple]
and [Favor] in the future, there are many stealing effects and it is
important to distinguish them properly.

In the alerting department, a lot of reduplication of code has been
minimised. Note that the test outputs still produce sane alerts, so we
know that these refactors are okay.

Miscellaneously, [Variants_helper] has been introduced since the pattern
of [Variants_helper.accumulate_without_args] and
[Variants_helper.accumulate_with_args] is used in 2 occasions of
[all_mocked]. No testing is done as these functions are straightforward
enough. They also clearly work when employed in tests.