Stop transfers between the same user, i.e. receiver = target
This resolves a bug where playing a [Double] or a [Triple] targeted at
yourself is allowed by the server. This is done by making
[transfer_random_card] validate by checking for [receiver] = [target].

However, addressing [transfer_random_card] only fixed [Double] as
[Triple] did not use the same logic - it did not need the randomness.
Hence, a small refactor is done to also expose
[Player_hands.transfer_card] for [Triple]. Then, [transfer_random_card]
inherits the logic from [transfer_card]. This way, when processing [Triple],
validation is also done, returning an error if [receiver] = [target].

Hence, tests are modified accordingly as well.