Add [Attack]
The [Attack] card is a Power card from the original game that skips your
turn immediately and gives the next player 2 turns. This is similar to
[Skip], but stronger. It also has a peculiar interaction with stacking -
if the recipient of an [Attack] also plays an [Attack], their 2 turns is
transferred to the next player, and the next player must take 2 + 2 = 4
turns.

Hence, when exposing the new function [give_all_turns_by_attack], a
strange-looking [if turns = 1 then ...] must be done, as the scaling
function is not linear (1 -> 2, 2 -> 4, 3 -> 5, 4 -> 6, ...).

Otherwise, [Turn_order] receives the biggest changes in this commit. A
new unexposed submodule [Player_and_turns] must be created as a player
may take multiple turns now, and this is reflected in [type t].
[pass_turn] and [give_all_turns_by_attack] are given greater scrutiny
and tests have been added for these 2 functions. A submodule
[Nonempty_list] is also created just to expose the [append'] function,
which is used in both of these functions. Hence, this refactor was
required to avoid duplication of code.

Otherwise, fairly straightforward changes such as adding [Power Attack]
is done and compilation errors that propogate are fixed accordingly. For
instance, [Outcome.Attacked] and [Next_step.Give_turns_via_attacking],
as well as the alert messages that correspond with these.