Implement [Favor]
The [Favor] card from the original game is a power card that allows you
to name another player. Then, that player must choose a card from their
hand and give it to you. This commit adds that card to the gameplay
loop.

In [card.ml], [Power] is broken down into [Targeted] and [Targetless].
This is because [Favor] requires a target to play unlike all
[Targetless] cards like [Skip]. This split is meaningful as it allows us
to split the parsing into 2 different well-defined cases, where [Favor]
can be played by typing "Favor@TARGET".

Hence, [Draw_or_play] now has [Play] broken down to [Play_targetless of
Card.Power.Targetless.t] and [Play_targeted of Card.Power.Targeted.t *
Player_name.t].

Accordingly, a new RPC, [Get_card_to_give], is introduced. A player
receives the [Hand.t] and picks a [Card.t] to return. All files such as
[logic.ml] and [interaction.ml] are thus modified to accomodate for this
new RPC.

[Outcome]s and [Next_step]s such as [Favored], [Receive_card_from] and
[Received_card_from] are added to cope with this new addition of
[Favor]. Note that when [Favor] is played, the current player remains as
the current player while the code branches into asking for the target
for a card. This means that [Favored], [Receive_card_from] and
[Received_card_from] only need to encode the target that it is coming
from - the current player is still the person who played [Favor] and
maintained.

An interesting scenario that was caught is the fact that the
[Player_name.t] provided upon playing [Favor] needs to validated before
returning [Favored target]. This is because during the [handle] stage,
any and all errors that are caught must be immediately returned so
re-prompting can be done. If it is only discovered that the target is
invalid at the [Receive_card_from] stage, it is too late and the code
will inevitably crash. This is why [Player_hands.is_playing] is a new
function that must be created.