Create [Callbacks] module
The [advance] functions in [Game_state] is starting to accumulate too
many parameters - 6 of them. This is a good sign that a new layer of
abstraction needs to be created.

[Callbacks] handles all of these parameters in a record. Upon creating
this, the functions in [Interaction] are appropriate enough to be
brought into [Callbacks] as they are very closely tied to the module.
This is merely a move of functions and no logic is changed here.