Remove duplicated code in [to_specialised_alert]
The code involved in [to_specialised_alert] has quite a bit of
reduplication and is very difficult to read. This commit tries to reduce
that.

The function is split out into [target], which returns the [target] of
[t] if it exists, as well as [rename_target_to_you]. This makes the code
more self-documenting.

The behaviour of this function is not affected, as supported by the
tests still passing.