Reduce nesting in [Action.of_string]
The nested regex matching and parsing gets very difficult to maintain.
Not only is there a lot of indentation, the order of which regex is
attempted for parsing is also not obvious, spanning multiple screens of
code. It is also not caught by the compiler for a new parser to be added
if more [Action]s are created.

This commit uses [Variants] to ensure the totality of the variant in
[of_string], while also providing a more reasonable error message upon a
failed parse.