Generate [opam] file with [dune-project] to manage dependencies
The [opam] file is most useful in managing dependencies. This commit
removes the wrongly added [opam] file in [.gitignore]. Then, the
documentation for installation is updated to use [opam install
--deps-only .]. [dune-project] is updated to include all the minimum
information as well as the dependencies for [opam] to be used
without errors.

The versions of the dependencies are stated exactly as the version I
have been using to compile the project. [async], [core] and [re2] are
stated with [=] for the version number as these are Jane Street related
libraries. Jane Street does not fear breaking back-compatibility when
upgrading versions, so using the exact version number is the safest.
[dune] does not have a version number in [depends] as the first line,
which states [lang dune 3.16], already dictates its version.

The newly added [public_names] in [/bin/dune] is required for dune to
compile successfully. They are otherwise insignificant.