Add more initial load details
This commit adds more information at the start of the game. The turn
order is not broadcast to players before this commit, which is not
enough information. The players should know who their opponents are and
what the turn order is upon starting.

This means that [on_initial_load] must now also take a [Turn_order.t] to
know what to send to players.

The string representation of [Turn_order] is encapsulated in
[Turn_order.to_string] and tests are added accordingly for this.